requirements.txt
requests
python-dotenv
pandas
fastapi
uvicorn
openpyxl
pytest
